HAR/requirements.txt
PyYAML==5.3
pandas==1.0.1
numpy==1.18.1
matplotlib==3.1.3
tensorflow-gpu
tensorboard==2.3.0
seaborn==0.10.0
scipy==1.4.1
scikit-learn==0.22.1
bunch==1.0.1
pydot==1.4.1
pymongo
keras-tcn
